pyproject.toml
[project]
name = "venv"
version = "0.1.0"
description = ""
authors = [
    {name = "Ben Parfitt",email = "ben@diligently.ai"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "discord-py (>=2.4.0,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
